pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qpcr-quality-control"
version = "0.1.0"
description = "Deterministic qPCR QC pipeline scaffold with HMM-style state calling."
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["PyYAML>=6.0"]

[project.optional-dependencies]
dev = ["pytest>=9.0"]
workflow = ["snakemake>=8,<9"]

[project.scripts]
qpcr-quality-control = "src.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q -p no:cacheprovider --basetemp=.pytest_tmp"
